requirements.txt
argostranslate==1.9.1
dbus-python==1.3.2
pyperclip==1.8.2
